* Converted math cache from doubles to floats * Fixed major CPU hog: idle was being called many times per frame!  This   should fix leaf speed. * Improved leaf coloration on some video cards.  Probably made worse on others, though.

+++ eye_candy/math_cache.cpp
@@ -8,7 +8,7 @@
 
 #include "math_cache.h"
 
-const double PI = 3.141592654;
+const float PI = 3.141592654;
 
 namespace ec
 {
